Refactor test_prompt.py by segregating tests into separate modules
